+++ ppracer/src/render_util.h
@@ -32,15 +32,15 @@ class Player;
 
 #define NEAR_CLIP_DIST 0.1
 
-void reshape( int w, int h, int multiscreen=-1);
+void reshape(int w, int h, int multiscreen=-1);
 
 void clear_rendering_context();
 
 void set_material(const ppogl::Color& diffuse, const ppogl::Color& specular,
-		   const double specular_exp);
+		   const float specular_exp);
 
 void draw_billboard(const Player& plyr,
-		     const ppogl::Vec3d& center_pt, double width, double height, 
-		     const ppogl::Vec2d& min_tex_coord, const ppogl::Vec2d& max_tex_coord );
+		     const ppogl::Vec3d& center_pt, float width, float height, 
+		     const ppogl::Vec2d& min_tex_coord, const ppogl::Vec2d& max_tex_coord);
 
 #endif // _RENDER_UTIL_H_
