+++ ppracer/src/view.h
@@ -25,7 +25,7 @@
 #include "player.h"
 
 void set_view_mode(Player& plyr, ViewMode mode);
-void update_view(Player& plyr, double dt);
+void update_view(Player& plyr, float dt);
 void setup_view_matrix(Player& plyr);
 
 #endif // _VIEWMODE_H_
